types: update types for 32/64-bit platforms

+++ userspace/tests/ipc/ping.c
@@ -11,17 +11,17 @@ const size_t data_size = 43;
 
 static void thread_main(void *arg)
 {
-    fd_t client_fd = (fd_t) (long) arg;
+    fd_t client_fd = (fd_t) arg;
 
     char buf[150];
-    size_t sz = sprintf(buf, "Welcome to the server, client fd %d!", client_fd);
+    size_t sz = sprintf(buf, "Welcome to the server, client fd %ld!", client_fd);
 
     size_t written = syscall_io_write(client_fd, buf, sz, 0);
     if (written != sz)
         printf("Server: failed to write to ipc channel\n");
 
     size_t read_size = syscall_io_read(client_fd, buf, 150, 0);
-    printf("Server: Received '%.*s' from client %d\n", (int) read_size, buf, client_fd);
+    printf("Server: Received '%.*s' from client %ld\n", (int) read_size, buf, client_fd);
 
     syscall_io_close(client_fd);
 }
@@ -57,8 +57,8 @@ int main(int argc, char **argv)
             return -1;
         }
 
-        printf("Server: Accepted fd %d\n", client_fd);
-        start_thread("child", thread_main, (void *) (long) client_fd);
+        printf("Server: Accepted fd %ld\n", client_fd);
+        start_thread("child", thread_main, (void *) client_fd);
     }
 
     return 0;
